Add the page visibility API support on sounds.
The sound manager is the one that is called when the viewer is
paused/resumed and thus all currently playing sounds are paused/resumed.
This change adds a new state to a sound: resumed. This state tells the
sound manager if the sound needs to be resumed after being paused.