Massive rework to support transition
Remove every rendering object from the Scene object and add new objects SceneRenderer and SceneRendererPool owned by the Renderer.
SceneRenderer owns the viewportManager, objectRenderer and the renderTarget of the scene.
Predefinition of a activeScene concept added to the activeViewport. Should evolve in the future.
Fix controllers with current changes.

Render routines cascade and pass the render target.

Everthing in this file is temporary and subject to discussions before integration.