VR compat
Enable access to VR with THREE WebVRManager.

Viewer has now its own VR flag, used by all the code to know when VR is turned ON/OFF.

WebGLRenderingContext attribute preserveDrawingBuffer should be set to true if user wants to mirror headset display to the screen.

As Forge uses many render targets, we need to distinguish renders to target (vr enabled) and to screen (vr disabled) to prevent from splitting things twice. This is done by enabling webGL Renderer vr.enabled flag before rendering background and objetcs to the scene texture and disabling it in the screen renderer at the start of the render call.

Submitting frame to the headset should as opposite be done only by the screen renderer, and only once the scenes have all been renderered. We use the scene callback onAfterRender to toggle the vr.enable flag in order to obtain this behaviour.